Finallizing support for API11 and above, a.k.a Android 3.0+
To be able to set immersive mode and other related fullscreen features,
it gets too tiresome below API 11. I am doing this because in the
CanvasActivity class, function named setImmersiveMode() the second line
calls upon the function "getSystemVisibility()" which belongs to API11
and I am unable to find a suitable alternative to it in API10 and below.