Trying to seperate Logging
This is just a first attempt to seperate logging constants and functions
from the code. I think this will require a lot of studying on my part to
be able to first understand the concept of logging in Android with it's
different priority levels and sepearting it into a library.