Trying to seperate tools - attempt ONE
Okay, so the intention is to make the tools modular. Currently I am
trying to create a base class named "Tool" which will point to the
objects of concrete classes derived from Tool. Each kind of tool will be
responsible for handling the touchevents by themselves, also to manage
and modify the Paint and Path objects.

For the uber geeks... this is supposed to be something akin to "Bridge"
design pattern.