Implementing MVC
While trying to seperate tools, I realized it would be very difficult to
keep the tools seperate and modular from the main activity or view class
if we don't loosen up the couping between the components. So, this is
the first step. This code refactoring is incomplete and has introduced
major amount of bugs back in.