Implemented Observer Pattern
Final intent is to decouple the model from controller (activity)
completely, and the view should be notified by the model itself whenever
it changes so that view can update itself.
I will also try to implement Protection Proxy over the model so that the
view is only able to read the model not mess with it in any way.